updated my file

+++ src/cart_items/service.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import Session
 from src.cart_items.models import CartItemCreate
 from src.entities.carts import CartItem
-from src.entities.users import User
+from src.entities.users import User, UserRole
 from src.entities.products import Product
 from fastapi import HTTPException, Depends, Response, status
 from src.auth.service import get_current_user
@@ -12,7 +12,8 @@ def add_to_cart(item: CartItemCreate, current_user: User = Depends(get_current_u
     product = db.query(Product).filter(Product.id == item.product_id).first()
     if not product:
         raise HTTPException(status_code=404, detail="Product not found")
-    
+    if UserRole.CUSTOMER != current_user.role:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Insufficient Permission")
     existing_item = db.query(CartItem).filter(
         CartItem.user_id == current_user.id,
         CartItem.product_id == item.product_id
@@ -33,21 +34,27 @@ def add_to_cart(item: CartItemCreate, current_user: User = Depends(get_current_u
 
 
 def get_cart(current_user: User = Depends(get_current_user), db: Session = Depends(get_db)):
-
-    return db.query(CartItem).filter(CartItem.user_id == current_user.id).all()
+    if UserRole.CUSTOMER != current_user.role:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Insufficient Permission")
+    cart = db.query(CartItem).filter(CartItem.user_id == current_user.id).all()
+    return cart
 
 
 
 def update_cart_item(item_id: int, quantity: int, current_user: User = Depends(get_current_user), 
                     db: Session = Depends(get_db)):
     cart_item = db.query(CartItem).filter(CartItem.id == item_id, CartItem.user_id == current_user.id).first()
+    
+    if UserRole.CUSTOMER != current_user.role:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Insufficient Permission")
+    cart_item.quantity = quantity
     if not cart_item:
         raise HTTPException(status_code=404, detail="Cart item not found")
     
-    cart_item.quantity = quantity
     db.commit()
     db.refresh(cart_item)
     
+
     return cart_item
 
 
@@ -55,7 +62,8 @@ def remove_from_cart(item_id: int, current_user: User = Depends(get_current_user
     cart_item = db.query(CartItem).filter(CartItem.id == item_id, CartItem.user_id == current_user.id).first()
     if not cart_item:
         raise HTTPException(status_code=404, detail="Cart item not found")
-    
+    if UserRole.CUSTOMER != current_user.role:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Insufficient Permission")
     db.delete(cart_item)
     db.commit()
